Botón para eliminar cuenta desde la vista de gestión de usuarios del admin y otras modificaciones

+++ cartera-tiuca/src/main/java/es/uca/iw/carteratiuca/services/ProyectoService.java
@@ -55,7 +55,8 @@ public class ProyectoService {
 
     public void delete(UUID id) {
         Proyecto proyecto = proyectoRepository.findById(id).get();
-        justificacionProyectoRepository.deleteById(proyecto.getJustificacion().getId());
+        if(proyecto.getJustificacion() != null)
+            justificacionProyectoRepository.deleteById(proyecto.getJustificacion().getId());
         proyectoRepository.deleteById(id);
     }
 }
